+++ inst/doc/spant-intro.R
@@ -1,8 +1,8 @@
-## ----setup, include = FALSE----------------------------------------------
+## ----setup, include = FALSE---------------------------------------------------
 knitr::opts_chunk$set(fig.width = 5, fig.height = 4, echo = TRUE)
 library(spant)
 
-## ----useage, message = FALSE---------------------------------------------
+## ----useage, message = FALSE--------------------------------------------------
 library(spant)
 fname <- system.file("extdata", "philips_spar_sdat_WS.SDAT", package = "spant")
 
@@ -15,13 +15,13 @@ print(mrs_data)
 # plot data in the frequency domain
 plot(mrs_data, xlim = c(5, 0.5))
 
-## ----processing, message = FALSE-----------------------------------------
+## ----processing, message = FALSE----------------------------------------------
 # apply water filter and align to tNAA resonance
 mrs_proc <- hsvd_filt(mrs_data)
 mrs_proc <- align(mrs_proc, 2.01)
 plot(mrs_proc, xlim = c(5, 0.5))
 
-## ----basis_sim, message = FALSE------------------------------------------
+## ----basis_sim, message = FALSE-----------------------------------------------
 # simulate a typical basis set for short TE brain analysis
 basis <- sim_basis_1h_brain_press(mrs_proc)
 
@@ -31,7 +31,7 @@ print(basis)
 # plot basis signals
 stackplot(basis, xlim = c(4, 0.5))
 
-## ----fitting, message = FALSE--------------------------------------------
+## ----fitting, message = FALSE-------------------------------------------------
 # perform VARPRO fitting to processed data
 fit_res <- fit_mrs(mrs_proc, basis)